Fixed turn advancement bug
