+++ frontend/src/pages/MagnateGame/LobbyManager.css
