+++ frontend/src/pages/MagnateGame/PlayArea/Game.css
